test: deduplicate storage setup
